modified:   db/borroTweet.go 	modified:   routers/borrarTweet.go

+++ db/borroTweet.go
@@ -2,6 +2,8 @@ package db
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"time"
 
 	"github.com/zeeno77/microblog/models"
@@ -9,64 +11,26 @@ import (
 )
 
 /*ModificoRegistro permite modificar el perfil del tweet*/
-//func BorroTweet(twt models.Tweet, ID string) (bool, error) {
-func BorroTweet(twt models.Tweet) (bool, error) {
+func BorroTweet(twt models.Tweet, ID string) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
 	defer cancel()
 
+	if twt.Userid.Hex() != ID {
+		println("Ese tweet no le pertenece")
+		return false, errors.New("not allowed")
+	}
 	db := MongoCN.Database(DBName)
 	col := db.Collection("tweet")
 
-	/* supongo que me van a enviar de a un campo a modificar a la vez,
-	   por eso me fijo si lo que viene tiene valor (largo mayor a cero)
-	   Creamos un mapa de interfaces para armar el registro de actualización a la bd
-	   Le pongo la información que hay que modificar*/
-	//	registro := make(map[string]interface{})
-
-	//	if len(twt.Nombre) > 0 {
-	//		registro["nombre"] = twt.Nombre
-	//	}
-	//	if len(twt.Apellidos) > 0 {
-	//		registro["apellidos"] = twt.Apellidos
-	//	}
-	//	registro["fechaNacimiento"] = twt.FechaNacimiento
-	//	if len(twt.Avatar) > 0 {
-	//		registro["avatar"] = twt.Avatar
-	//	}
-	//	if len(twt.Banner) > 0 {
-	//		registro["banner"] = twt.Banner
-	//	}
-	//	if len(twt.Biografia) > 0 {
-	//		registro["biografia"] = twt.Biografia
-	//	}
-	//	if len(twt.Ubicacion) > 0 {
-	//		registro["ubicacion"] = twt.Ubicacion
-	//	}
-	//	if len(twt.SitioWeb) > 0 {
-	//		registro["sitioWeb"] = twt.SitioWeb
-	//	}
-	//	updtString := bson.M{
-	//		"$set": registro,
-	//	}
-
-	//	if len(twt.userId) > 0 {
-	//		if ID == twt.userId {
-	//			println("Este tweet es mio")
-	//		} //si el tweet es mio
-	//	}
-
-	//objID, _ := primitive.ObjectIDFromHex(twt.ID)
-	// ahora tengo que realizar un filtro con el ID
-
-	println("******************")
-	println(twt.ID.Hex())
-	println("******************")
 	filtro := bson.M{"_id": bson.M{"$eq": twt.ID}}
-
+	//Do not delete, just print on screen
+	//deletion is working
 	_, err := col.DeleteOne(ctx, filtro)
 	if err != nil {
 		return false, err
 	}
+	fmt.Println("Tweet borrado con exito")
+
 	return true, nil
 
 }
